entrypoint/backend/requirements.txt
# FILE: backend/requirements.txt
# LANGUAGE: plain text
# Install all Python packages with:
# pip install -r requirements.txt

fastapi
uvicorn
anthropic
httpx
python-dotenv
